A possible way to fix the login
There are several ways to this, but essentially, you need to make sure that you're passing to UserAPIService.callAPI an object that has "username" and "password" fields. Since $scope.user is already such an object, which can just pass that, and remove a few lines of code that are no longer needed.